fix(analytics): cast processor_entry_ms in the inbound_ms backfill + add live latency check
The inbound_ms derivation used the processor_entry_ms parameter only in `IS NOT NULL` and arithmetic,
so Postgres could not infer its type and the exchange backfill failed at execution ("could not determine
data type of parameter $16"), dropping e2e_ms and inbound_ms for the row. The mock-level unit test
asserts the SQL string, not execution, so it did not catch this. Cast the parameter to bigint.

Adds tests/e2e/latency.spec.ts: a live end-to-end check that drives one tool-using turn and asserts the
latency_metrics row carries e2e_ms / total_ms / ttff_ms / model_ms > 0 - the value-level proof the
mock-level test cannot provide.

+++ backend/lambda/src/analytics-aurora/kinesis-archival.ts
@@ -983,8 +983,8 @@ export async function backfillFromUpdateEvents(
                 -- start). CROSS-CLOCK ($16 server-clock entry vs ex.user_message_at Chime); GREATEST
                 -- clamps the skew to >= 0. Approximate by design.
                 inbound_ms        = COALESCE(ex.inbound_ms,
-                                      CASE WHEN $16 IS NOT NULL
-                                           THEN GREATEST(0, ($16 - EXTRACT(EPOCH FROM ex.user_message_at) * 1000)::integer)
+                                      CASE WHEN $16::bigint IS NOT NULL
+                                           THEN GREATEST(0, ($16::bigint - (EXTRACT(EPOCH FROM ex.user_message_at) * 1000)::bigint)::integer)
                                       END)
            FROM messages am
           WHERE ex.agent_message_id = am.id
